backend/requirements.txt
python-dotenv==1.0.0
pydantic==2.13.4
pydantic-settings==2.14.1
langgraph==0.1.19
google-generativeai==0.3.0
requests==2.31.0
SQLAlchemy==2.0.23
faker==22.0.0
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.3
phonenumbers==8.13.0
rapidfuzz==3.5.2
streamlit==1.30.0
fastapi==0.109.0
uvicorn==0.27.0
twilio==9.0.0
APScheduler==3.10.4
python-jose==3.3.0
passlib==1.7.4
